% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_grand_average_erp.R
\name{compute_grand_average_erp}
\alias{compute_grand_average_erp}
\title{compute_grand_average_erp}
\usage{
compute_grand_average_erp(erpdata, electrode)
}
\arguments{
\item{erpdata}{Input Matrix with shape: subject x condition x time x electrodes.}

\item{electrode}{A numeric value that specifies the number of electrode that should be plotted}
}
\value{
A matrix that contains the grand average ERP for all conditions in a selected electrode. The matrix will be of shape condition x time.
}
\description{
The function \emph{grand_average_erp} computes the grand average  ERP signal for all conditions and a given electrode.
}
\details{
For more details, see Wikipedie page: \href{https://sccn.ucsd.edu/wiki/I.6:_Data_Averaging}{article}
}
\seealso{
Add example of similar functions [still to be done].
}
