% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_average_ersp.R
\name{compute_grand_average_ersp}
\alias{compute_grand_average_ersp}
\title{compute_grand_average_ersp}
\usage{
compute_grand_average_ersp(my_erspdata)
}
\arguments{
\item{my_erspdata}{Input Matrix with shape: subject ID x condition x timex frequency (from only one electrode).}
}
\value{
A matrix that contains the grand average ERSP for all conditions and all frequencies of one electrode. The matrix will be of shape condition x time x frequency.
}
\description{
The function \emph{grand_average_ersp} computes the grand average  ERSP signal for all conditions and all frequencies from one electrode.
}
\details{
For more details, see Wikipedie page: \href{https://sccn.ucsd.edu/wiki/Chapter_11:_Time/Frequency_decomposition}{article}
}
