% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ind_average_erp.R
\name{compute_ind_average_erp}
\alias{compute_ind_average_erp}
\title{compute_ind_average_erp}
\usage{
compute_ind_average_erp(erpdata, electrode, subject)
}
\arguments{
\item{erpdata}{Input Matrix with shape: subject x condition x time x electrodes.}

\item{electrode}{A character that specifies the electrode that should be plotted.}

\item{subject}{A character or a vector of characters that contains the number of subject (e.g. 'SO1')}
}
\value{
A matrix that contains the single-subject ERP for all conditions in a selected electrode. The matrix will be of shape electrode x condition x subject.
}
\description{
The function \emph{ind_average_erp} computes the single subject ERP signal for all conditions and a given electrode.
}
\details{
For more details, see Wikipedie page: \href{https://sccn.ucsd.edu/wiki/I.6:_Data_Averaging}{article}
}
