% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_erp_ersp.R
\name{correlate_erp_ersp}
\alias{correlate_erp_ersp}
\title{correlate_erp_ersp}
\usage{
correlate_erp_ersp(erpdata, electrode, time_erp, my_erspdata, frequency)
}
\arguments{
\item{erpdata}{Input Matrix with shape: subject ID x condition x time x electrode}

\item{electrode}{A character that specifies the number of electrode that should be used for ERP calculation.}

\item{time_erp}{A numeric value that specifies the time of the ERP (e.g. 20) in ms}

\item{my_erspdata}{Input Matrix with shape: subject ID x condition x time x frequency (from only one electrode).}

\item{frequency}{A numeric value that specifies the desired frequency.}
}
\value{
A correlation matrix with the correlations between the specified erp and ersp for all conditions.

A correlation plot thatr illustrates the ERP / ERSP correlations for all conditions.
}
\description{
The function \emph{erp_ersp_cor} correlates a selected ERP and ERSP component.
}
