% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GA_Plot}
\alias{GA_Plot}
\title{GA_Plot}
\usage{
GA_Plot(my_erpdata, electrode = 6)
}
\arguments{
\item{my_erpdata}{Input Matrix with shape: subject x condition x time x electrodes.}

\item{electrode}{A numeric value that specifies the number of electrode that should be plotted}
}
\value{
A plot with the Grand Average.
}
\description{
The function \emph{GA_Plot} computes the grand average ERP signal for all conditions and plots it.
}
\details{
For more details, see Wikipedie page: \href{https://sccn.ucsd.edu/wiki/I.6:_Data_Averaging}{article}
}
