% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grand_average_erp.R
\name{plot_grand_average_erp}
\alias{plot_grand_average_erp}
\title{plot_grand_average_erp}
\usage{
plot_grand_average_erp(my_erpdata, grand_average_erp_data)
}
\arguments{
\item{my_erpdata}{Input Matrix with shape: subject x condition x time x electrodes.}

\item{grand_average_erp_data}{Input Matrix that contains the single-subject ERP for all conditions in a selected electrode. The matrix should be of shape condition x time.}
}
\value{
A plot with the grand average ERP signal.
}
\description{
The function \emph{grand_average_erp_plot} plots the grand average ERP.
}
