% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grand_average_ersp.R
\name{plot_grand_average_ersp}
\alias{plot_grand_average_ersp}
\title{plot_grand_average_ersp}
\usage{
plot_grand_average_ersp(my_erspdata, grand_average_ersp_data, n_cond)
}
\arguments{
\item{my_erspdata}{Input Matrix with shape: subject ID x condition x timex frequency (from only one electrode).}

\item{grand_average_ersp_data}{A matrix that contains the grand average ERSP for all conditions and all frequencies of one electrode.
The matrix will be of shape condition x time x frequency.}

\item{ncond}{A numeric value that specifies which condition should be plotted (e.g. 1 for condition 1)}
}
\value{
A time - frequency plot with the grand average ERSP signal.
}
\description{
The function \emph{grand_average_ersp_plot} plots the grand average ERSP as a time frequency plot.
}
