% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind_average_erp.R
\name{plot_ind_average_erp}
\alias{plot_ind_average_erp}
\title{plot_ind_average_erp}
\usage{
plot_ind_average_erp(erpdata, ind_average_erp_data)
}
\arguments{
\item{erpdata}{Input Matrix with shape: subject x condition x time x electrodes.}

\item{ind_average_erp_data}{Input Matrix that contains the single-subject ERP for all conditions in a selected electrode. The matrix should be of shape condition x subject.}
}
\value{
A plot with the sinlge subject - ERP.
}
\description{
The function \emph{ind_average_plot} plots the single-subject ERP.
}
\seealso{
Add example of similar functions [still to be done].
}
