% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_data.R
\name{reading_data}
\alias{reading_data}
\title{reading_data}
\usage{
reading_data(erpdata, electrode_names)
}
\arguments{
\item{erpdata}{Input Matrix with shape: subject ID x condition x time x electrodes}

\item{electrode_names}{Vector of characters for the respective electrode names}
}
\value{
A short description of the input data and potential concerns as warnings.
}
\description{
The function \emph{reading_data} checks whether the shape of the input matrix is correct
and adds the electrode names to the datafile. This function should be run before running other eegstats
functions. If there are indications that the shape of the input file is not correct, warning messages
with relevant instructions will appear. The function also prints a short description of the dataset.
}
